Use proxy strategy as the label for proxy_server_ready_backend_connections_total metric
